requirements.txt
python-telegram-bot==11.1.0
sqlalchemy==1.3.0
pytest==4.5.0
mock==3.0.5

# for the healthcheck mini-app
py-healthcheck==1.8.1
Flask==1.0.3
gevent==1.4.0

ipython[all]
ipdb
